refactor(events): split member scan service and memberships tests under file-length limit
ticket_service.py and test_me_subscriptions.py crossed the 1000-line cap
after Tasks 7/9/11 of the membership-passes work. Move MemberScanResult
and scan_member_code() into a new member_scan_service.py (companion to
ticket_service, imports check_in_ticket from it) and move the
/me/memberships test class into test_me_memberships.py. No behavior
change; consumers updated accordingly.
